can save meshes

+++ ScanMeasure_v2/GL_Extension/Angel.h
@@ -64,7 +64,7 @@ GLuint InitShader( const char* vertexShaderFile,
 const GLfloat  DivideByZeroTolerance = GLfloat(1.0e-07);
 
 //  Degrees-to-radians constant 
-const GLfloat  DegreesToRadians = M_PI / 180.0;
+const GLfloat  DegreesToRadians = (GLfloat)M_PI / 180;
 
 }  // namespace Angel
